feat: events now are stored in the DomainEvents singletone Added AggregateEvent inside DomainEvents to help aggregates managing their events.
DomainEvents now is an abstract class and thus cannot be instantiated and should neither be implemented.

Aggregates now have a `dispose` method that should be called after an domain event is registered for dispatch, so that it won't let any unused or useless domain events list in the memory (especially if the repositories fails to connect to the database and throw an error before dispatching the event, cleaning it from memory). More details about the method usage can be found in `src/core/aggregate.ts` jsDocs.

Added a try-finally block in every service that deals with domain events, calling the dispose method if the repository fails and thus not call the event dispatch method.